+++ pkg/randomdecimal/randomdecimal_test.go
@@ -43,9 +43,7 @@ func TestMaskingShouldReplaceSensitiveValueByRandomNumber(t *testing.T) {
 
 func TestFactoryShouldCreateAMask(t *testing.T) {
 	maskingConfig := model.Masking{Mask: model.MaskType{RandomDecimal: model.RandomDecimalType{Min: float64(0), Max: float64(10), Precision: 2}}}
-	conf, present, err := Factory(maskingConfig, 0)
-	waitedConf := NewMask(float64(0), float64(10), 2, 0)
-	assert.Equal(t, conf, waitedConf, "should be equal")
+	_, present, err := Factory(maskingConfig, 0)
 	assert.True(t, present, "should be true")
 	assert.Nil(t, err, "error should be nil")
 }